# packages/api/llm_utils/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chain Swarm API",
    "description": "API for accessing chain swarm",
    "version": "0.2.0"
  },
  "paths": {
...
    "/substrate/{network}/balance-transfers/address/{address}": {
      "get": {
        "tags": [
          "balance-transfers"
        ],
        "summary": "Get Address Transactions",
        "description": "Retrieves transaction history for a specific address on a blockchain network.\n\nThis endpoint provides a paginated list of transactions involving the specified address, either as sender or receiver. Optionally, you can filter transactions between the specified address and a target address.",
        "operationId": "get_address_transactions_substrate__network__balance_transfers_address__address__get",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain network identifier",
              "title": "Network"
            },
            "description": "The blockchain network identifier",
            "example": "torus"
          },
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain address to query",
              "title": "Address"
            },
            "description": "The blockchain address to query",
            "example": "5C4n8kb3mno7i8vQmqNgsQbwZozHvPyou8TAfZfZ7msTkS5f"
          },
          {
            "name": "target_address",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter transactions to/from this address",
              "title": "Target Address"
            },
            "description": "Filter transactions to/from this address"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number for pagination",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number for pagination"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of transactions per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Number of transactions per page"
          },
          {
            "name": "assets",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
              "title": "Assets"
            },
            "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
            "example": [
              "TOR"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Address not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/substrate/{network}/balance-transfers/volume-series": {
      "get": {
        "tags": [
          "balance-transfers"
        ],
        "summary": "Get Balance Transfers Volume Series",
        "description": "Retrieves balance transfers volume series data providing network-wide transfer activity metrics.\n\nThis endpoint provides paginated balance transfers volume series data showing network activity metrics like transaction counts per period, transfer volumes per period, active addresses per period, and other network activity indicators. Supports different time aggregations (4-hour, daily, weekly, monthly) and filtering by assets and time range.",
        "operationId": "get_balance_transfers_volume_series_substrate__network__balance_transfers_volume_series_get",
        "parameters": [
          {
            "name": "network",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The blockchain network identifier",
              "title": "Network"
            },
            "description": "The blockchain network identifier",
            "example": "torus"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number for pagination",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number for pagination"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Number of volume series records per page",
              "default": 20,
              "title": "Page Size"
            },
            "description": "Number of volume series records per page"
          },
          {
            "name": "assets",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
              "title": "Assets"
            },
            "description": "List of assets to filter by. Use ['all'] for all assets. Defaults to network's native asset.",
            "example": [
              "TOR"
            ]
          },
          {
            "name": "start_timestamp",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start timestamp in milliseconds (Unix timestamp)",
              "title": "Start Timestamp"
            },
            "description": "Start timestamp in milliseconds (Unix timestamp)",
            "example": 1742815440000
          },
          {
            "name": "end_timestamp",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End timestamp in milliseconds (Unix timestamp)",
              "title": "End Timestamp"
            },
            "description": "End timestamp in milliseconds (Unix timestamp)",
            "example": 1842815440000
          },
          {
            "name": "period_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(4hour|daily|weekly|monthly)$",
              "description": "Period type for aggregation",
              "default": "4hour",
              "title": "Period Type"
            },
            "description": "Period type for aggregation",
            "example": "4hour"
          }
        ],
        "responses": {
          "200": {
            "description": "Balance transfers volume series retrieved successfully",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "400": {
            "description": "Invalid period type specified"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}